proto: move runtime code into a module
