+++ bin/tests/integration/named_quic_tests.rs
@@ -65,7 +65,7 @@ fn test_example_quic_toml_startup() {
 
         // ipv4 should succeed
         let (mut client, bg) = io_loop.block_on(client).expect("client failed to connect");
-        hickory_proto::spawn_bg(&io_loop, bg);
+        hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
         query_a(&mut io_loop, &mut client);
