+++ crates/client/src/client/async_secure_client.rs
@@ -14,11 +14,11 @@ use futures_util::stream::Stream;
 use crate::client::AsyncClient;
 use crate::proto::error::ProtoError;
 use crate::proto::rr::dnssec::TrustAnchor;
+use crate::proto::runtime::TokioTime;
 use crate::proto::xfer::{
     DnsExchangeBackground, DnsHandle, DnsRequest, DnsRequestSender, DnsResponse,
 };
 use crate::proto::DnssecDnsHandle;
-use crate::proto::TokioTime;
 
 /// A DNSSEC Client implemented over futures-rs.
 ///
