+++ crates/proto/src/multicast/mdns_client_stream.rs
@@ -17,8 +17,9 @@ use futures_util::stream::{Stream, StreamExt, TryStreamExt};
 use crate::error::ProtoError;
 use crate::multicast::mdns_stream::{MDNS_IPV4, MDNS_IPV6};
 use crate::multicast::{MdnsQueryType, MdnsStream};
+use crate::runtime::TokioTime;
 use crate::xfer::{DnsClientStream, SerialMessage};
-use crate::{BufDnsStreamHandle, TokioTime};
+use crate::BufDnsStreamHandle;
 
 /// A UDP client stream of DNS binary packets
 #[must_use = "futures do nothing unless polled"]
