+++ crates/proto/src/native_tls/tests.rs
@@ -33,7 +33,7 @@ use tokio::runtime::Runtime;
 #[allow(unused)]
 use crate::native_tls::{TlsStream, TlsStreamBuilder};
 use crate::xfer::SerialMessage;
-use crate::{iocompat::AsyncIoTokioAsStd, DnsStreamHandle};
+use crate::{runtime::iocompat::AsyncIoTokioAsStd, DnsStreamHandle};
 
 // this fails on linux for some reason. It appears that a buffer somewhere is dirty
 //  and subsequent reads of a message buffer reads the wrong length. It works for 2 iterations
