+++ tests/integration-tests/src/mock_client.rs
@@ -19,10 +19,10 @@ use hickory_proto::error::ProtoError;
 use hickory_proto::op::{Message, Query};
 use hickory_proto::rr::rdata::{CNAME, NS, SOA};
 use hickory_proto::rr::{Name, RData, Record};
+use hickory_proto::runtime::TokioTime;
 use hickory_proto::tcp::DnsTcpStream;
 use hickory_proto::udp::DnsUdpSocket;
 use hickory_proto::xfer::{DnsHandle, DnsRequest, DnsResponse};
-use hickory_proto::TokioTime;
 use hickory_resolver::config::{NameServerConfig, ResolverOpts};
 use hickory_resolver::name_server::{ConnectionProvider, RuntimeProvider};
 use hickory_resolver::TokioHandle;
