+++ tests/integration-tests/tests/integration/client_future_tests.rs
@@ -19,14 +19,16 @@ use hickory_client::{
     client::{AsyncClient, ClientHandle},
     error::ClientErrorKind,
 };
+use hickory_integration::{
+    example_authority::create_example, NeverReturnsClientStream, TestClientStream,
+};
 #[cfg(feature = "dnssec")]
 use hickory_proto::rr::{dnssec::SigSigner, Record};
+#[cfg(all(feature = "dnssec", feature = "sqlite"))]
+use hickory_proto::runtime::TokioTime;
 #[cfg(feature = "dnssec")]
 use hickory_proto::xfer::{DnsExchangeBackground, DnsMultiplexer};
-#[cfg(all(feature = "dnssec", feature = "sqlite"))]
-use hickory_proto::TokioTime;
 use hickory_proto::{
-    iocompat::AsyncIoTokioAsStd,
     op::{Edns, Message, MessageType, OpCode, Query, ResponseCode},
     rr::{
         rdata::{
@@ -35,16 +37,13 @@ use hickory_proto::{
         },
         DNSClass, Name, RData, RecordSet, RecordType,
     },
+    runtime::iocompat::AsyncIoTokioAsStd,
     tcp::TcpClientStream,
     udp::UdpClientStream,
     xfer::FirstAnswer,
     DnsHandle,
 };
 use hickory_server::authority::{Authority, Catalog};
-
-use hickory_integration::{
-    example_authority::create_example, NeverReturnsClientStream, TestClientStream,
-};
 
 #[test]
 fn test_query_nonet() {
@@ -58,7 +57,7 @@ fn test_query_nonet() {
     let (stream, sender) = TestClientStream::new(Arc::new(StdMutex::new(catalog)));
     let client = AsyncClient::new(stream, sender, None);
     let (mut client, bg) = io_loop.block_on(client).expect("client failed to connect");
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     io_loop.block_on(test_query(&mut client));
     io_loop.block_on(test_query(&mut client));
@@ -71,7 +70,7 @@ fn test_query_udp_ipv4() {
     let stream = UdpClientStream::<TokioUdpSocket>::new(addr);
     let client = AsyncClient::connect(stream);
     let (mut client, bg) = io_loop.block_on(client).expect("client failed to connect");
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // TODO: timeouts on these requests so that the test doesn't hang
     io_loop.block_on(test_query(&mut client));
@@ -91,7 +90,7 @@ fn test_query_udp_ipv6() {
     let stream = UdpClientStream::<TokioUdpSocket>::new(addr);
     let client = AsyncClient::connect(stream);
     let (mut client, bg) = io_loop.block_on(client).expect("client failed to connect");
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // TODO: timeouts on these requests so that the test doesn't hang
     io_loop.block_on(test_query(&mut client));
@@ -106,7 +105,7 @@ fn test_query_tcp_ipv4() {
     let (stream, sender) = TcpClientStream::<AsyncIoTokioAsStd<TokioTcpStream>>::new(addr);
     let client = AsyncClient::new(stream, sender, None);
     let (mut client, bg) = io_loop.block_on(client).expect("client failed to connect");
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // TODO: timeouts on these requests so that the test doesn't hang
     io_loop.block_on(test_query(&mut client));
@@ -125,7 +124,7 @@ fn test_query_tcp_ipv6() {
     let (stream, sender) = TcpClientStream::<AsyncIoTokioAsStd<TokioTcpStream>>::new(addr);
     let client = AsyncClient::new(stream, sender, None);
     let (mut client, bg) = io_loop.block_on(client).expect("client failed to connect");
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // TODO: timeouts on these requests so that the test doesn't hang
     io_loop.block_on(test_query(&mut client));
@@ -161,7 +160,7 @@ fn test_query_https() {
             .build::<AsyncIoTokioAsStd<TokioTcpStream>>(addr, "cloudflare-dns.com".to_string()),
     );
     let (mut client, bg) = io_loop.block_on(client).expect("client failed to connect");
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // TODO: timeouts on these requests so that the test doesn't hang
     io_loop.block_on(test_query(&mut client));
@@ -266,7 +265,7 @@ fn test_notify() {
     let (stream, sender) = TestClientStream::new(Arc::new(StdMutex::new(catalog)));
     let client = AsyncClient::new(stream, sender, None);
     let (mut client, bg) = io_loop.block_on(client).expect("client failed to connect");
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     let name = Name::from_str("ping.example.com").unwrap();
 
@@ -337,7 +336,7 @@ async fn create_sig0_ready_client() -> (
 fn test_create() {
     let io_loop = Runtime::new().unwrap();
     let ((mut client, bg), origin) = io_loop.block_on(create_sig0_ready_client());
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // create a record
     let record = Record::from_rdata(
@@ -383,7 +382,7 @@ fn test_create() {
 fn test_create_multi() {
     let io_loop = Runtime::new().unwrap();
     let ((mut client, bg), origin) = io_loop.block_on(create_sig0_ready_client());
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // create a record
     let record = Record::from_rdata(
@@ -440,7 +439,7 @@ fn test_create_multi() {
 fn test_append() {
     let io_loop = Runtime::new().unwrap();
     let ((mut client, bg), origin) = io_loop.block_on(create_sig0_ready_client());
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // append a record
     let record = Record::from_rdata(
@@ -518,7 +517,7 @@ fn test_append() {
 fn test_append_multi() {
     let io_loop = Runtime::new().unwrap();
     let ((mut client, bg), origin) = io_loop.block_on(create_sig0_ready_client());
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // append a record
     let record = Record::from_rdata(
@@ -603,7 +602,7 @@ fn test_append_multi() {
 fn test_compare_and_swap() {
     let io_loop = Runtime::new().unwrap();
     let ((mut client, bg), origin) = io_loop.block_on(create_sig0_ready_client());
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // create a record
     let record = Record::from_rdata(
@@ -659,7 +658,7 @@ fn test_compare_and_swap() {
 fn test_compare_and_swap_multi() {
     let io_loop = Runtime::new().unwrap();
     let ((mut client, bg), origin) = io_loop.block_on(create_sig0_ready_client());
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // create a record
     let mut current = RecordSet::with_ttl(
@@ -725,7 +724,7 @@ fn test_compare_and_swap_multi() {
 fn test_delete_by_rdata() {
     let io_loop = Runtime::new().unwrap();
     let ((mut client, bg), origin) = io_loop.block_on(create_sig0_ready_client());
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // append a record
     let record1 = Record::from_rdata(
@@ -776,7 +775,7 @@ fn test_delete_by_rdata() {
 fn test_delete_by_rdata_multi() {
     let io_loop = Runtime::new().unwrap();
     let ((mut client, bg), origin) = io_loop.block_on(create_sig0_ready_client());
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // append a record
     let mut rrset = RecordSet::with_ttl(
@@ -853,7 +852,7 @@ fn test_delete_by_rdata_multi() {
 fn test_delete_rrset() {
     let io_loop = Runtime::new().unwrap();
     let ((mut client, bg), origin) = io_loop.block_on(create_sig0_ready_client());
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // append a record
     let mut record = Record::from_rdata(
@@ -904,7 +903,7 @@ fn test_delete_all() {
 
     let io_loop = Runtime::new().unwrap();
     let ((mut client, bg), origin) = io_loop.block_on(create_sig0_ready_client());
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     // append a record
     let mut record = Record::from_rdata(
@@ -984,7 +983,7 @@ fn test_timeout_query_nonet() {
     let client =
         AsyncClient::with_timeout(stream, sender, std::time::Duration::from_millis(1), None);
     let (client, bg) = io_loop.block_on(client).expect("client failed to connect");
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     test_timeout_query(client, io_loop);
 }
@@ -1005,7 +1004,7 @@ fn test_timeout_query_udp() {
         UdpClientStream::<TokioUdpSocket>::with_timeout(addr, std::time::Duration::from_millis(1));
     let client = AsyncClient::connect(stream);
     let (client, bg) = io_loop.block_on(client).expect("client failed to connect");
-    hickory_proto::spawn_bg(&io_loop, bg);
+    hickory_proto::runtime::spawn_bg(&io_loop, bg);
 
     test_timeout_query(client, io_loop);
 }
